refactor(shm): move UnifiedHeader + layout math into core::header
The on-mmap header is structurally pure: magic, version, slot counters,
manifest fingerprint, generation, padding — every field is a generic
integer / atomic / byte buffer with no business interpretation. Pull
UnifiedHeader, UNIFIED_MAGIC, UNIFIED_VERSION, DEFAULT_MAX_SLOTS,
slot_offset(), and calculate_file_size() out of unified_shm.rs into a
new core::header module. The size==64 invariant assert moves with them.

unified_shm.rs now re-exports these from core::header to keep the
existing `unified_shm::*` import paths working. The `routing_hash`
field name is retained for now — what the hash fingerprints
(ChannelPointCounts) is a business concern; the header itself is just
a u64 atomic that comsrv and modsrv agree on by value. A future
rename to `manifest_hash` is purely cosmetic.

After this commit, the SHM binary contract (header layout + slot
offset arithmetic) lives entirely in core::, alongside core::slot
(PointSlot/seqlock), core::bitmap (allocator), and core::config
(paths/constants/time). Promotion of core/ to a standalone crate now
only requires the Writer/Reader split — no further header surgery.
